🐛 Fix: Corrige erros em ConsultaDAO e MedicoDAO
Corrige a querySQL em marcarConsulta e inserirMedico

+++ src/main/java/Arquitetura/Model/Consulta.java
@@ -50,7 +50,7 @@ public class Consulta {
         return idConsulta;
     }
 
-    public void setIdConsulta(int idConsulta) {
+    public void setIdConsulta(long idConsulta) {
         this.idConsulta = idConsulta;
     }
